Implement C++11 variadic templates based actor.
