+++ include/boost/phoenix/core/detail/cpp03/actor_operator.hpp
@@ -18,7 +18,7 @@
 #include <boost/phoenix/support/iterate.hpp>
 
 #if defined(__WAVE__) && defined(BOOST_PHOENIX_CREATE_PREPROCESSED_FILES)
-#pragma wave option(preserve: 2, line: 0, output: "preprocessed/actor_operator_" BOOST_PHOENIX_LIMIT_STR ".hpp")
+#pragma wave option(preserve: 2, line: 0, output: "../preprocessed/actor_operator_" BOOST_PHOENIX_LIMIT_STR ".hpp")
 #endif
 /*==============================================================================
     Copyright (c) 2005-2010 Joel de Guzman
@@ -38,7 +38,7 @@
 
 #define BOOST_PHOENIX_ITERATION_PARAMS                                          \
         (3, (1, BOOST_PHOENIX_ACTOR_LIMIT,                                      \
-        <boost/phoenix/core/detail/actor_operator.hpp>))
+        <boost/phoenix/core/detail/cpp03/actor_operator.hpp>))
 #include BOOST_PHOENIX_ITERATE()
 
 #undef M0
