+++ include/boost/phoenix/core/detail/cpp03/actor_result_of.hpp
@@ -18,7 +18,7 @@
 #include <boost/phoenix/support/iterate.hpp>
 
 #if defined(__WAVE__) && defined(BOOST_PHOENIX_CREATE_PREPROCESSED_FILES)
-#pragma wave option(preserve: 2, line: 0, output: "preprocessed/actor_result_of_" BOOST_PHOENIX_LIMIT_STR ".hpp")
+#pragma wave option(preserve: 2, line: 0, output: "../preprocessed/actor_result_of_" BOOST_PHOENIX_LIMIT_STR ".hpp")
 #endif
 
 /*==============================================================================
@@ -33,47 +33,15 @@
 #pragma wave option(preserve: 1)
 #endif
 
-    namespace result_of
-    {
         template <typename Expr
             , BOOST_PHOENIX_typename_A_void(BOOST_PHOENIX_ACTOR_LIMIT)
             , typename Dummy = void>
         struct actor;
 
-        template <typename Expr>
-        struct nullary_actor_result
-        {
-            typedef
-                typename boost::phoenix::evaluator::impl<
-                    Expr const&
-                  , vector2<
-                        vector1<const ::boost::phoenix::actor<Expr> *> &
-                      , default_actions
-                    > const &
-                  , proto::empty_env
-                >::result_type
-                type;
-        };
-
-        template <typename Expr>
-        struct actor<Expr>
-        {
-            typedef
-                // avoid calling result_of::actor when this is false
-                typename mpl::eval_if_c<
-                    result_of::is_nullary<Expr>::value
-                  , nullary_actor_result<Expr>
-                  , mpl::identity<detail::error_expecting_arguments>
-                >::type
-            type;
-        };
-
 #define BOOST_PHOENIX_ITERATION_PARAMS                                          \
     (3, (1, BOOST_PHOENIX_ACTOR_LIMIT,                                          \
-    <boost/phoenix/core/detail/actor_result_of.hpp>))
+    <boost/phoenix/core/detail/cpp03/actor_result_of.hpp>))
 #include BOOST_PHOENIX_ITERATE()
-
-    }
 
 #if defined(__WAVE__) && defined(BOOST_PHOENIX_CREATE_PREPROCESSED_FILES)
 #pragma wave option(output: null)
