+++ include/boost/phoenix/core/detail/preprocessed/assign.hpp
@@ -5,19 +5,19 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ==============================================================================*/
 
-#if !defined(BOOST_PHOENIX_PREPROCESSED_CORE_ACTOR_HPP)
-#define BOOST_PHOENIX_PREPROCESSED_CORE_ACTOR_HPP
+#if !defined(BOOST_PHOENIX_PREPROCESSED_CORE_ASSIGN_HPP)
+#define BOOST_PHOENIX_PREPROCESSED_CORE_ASSIGN_HPP
 
 #if BOOST_PHOENIX_LIMIT <= 10
-#include <boost/phoenix/core/preprocessed/actor_10.hpp>
+#include <boost/phoenix/core/detail/preprocessed/assign_10.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 20
-#include <boost/phoenix/core/preprocessed/actor_20.hpp>
+#include <boost/phoenix/core/detail/preprocessed/assign_20.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 30
-#include <boost/phoenix/core/preprocessed/actor_30.hpp>
+#include <boost/phoenix/core/detail/preprocessed/assign_30.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 40
-#include <boost/phoenix/core/preprocessed/actor_40.hpp>
+#include <boost/phoenix/core/detail/preprocessed/assign_40.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 50
-#include <boost/phoenix/core/preprocessed/actor_50.hpp>
+#include <boost/phoenix/core/detail/preprocessed/assign_50.hpp>
 #else
 #error "BOOST_PHOENIX_LIMIT out of bounds for preprocessed headers"
 #endif
